Support of 64-bit systems. Memory leak fixed.
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@73998 9d14d920-75d5-0310-8bd7-a52572981aee

+++ lib/tapi/tapi_sniffer.c
@@ -259,7 +259,7 @@ tapi_sniffer_add_mult(const char *ta, const char *iface, const char *name,
     tapi_sniffer_id      *newsnid = NULL;
     tapi_sniff_list_s    *newsn_l = NULL;
     cfg_handle           *handles;
-    int                   n_handles;
+    unsigned              n_handles;
     int                   rc;
     int                   i;
     char                 *newiface;
@@ -270,7 +270,7 @@ tapi_sniffer_add_mult(const char *ta, const char *iface, const char *name,
                                   "/agent:%s/interface:*", ta);
         if (rc != 0)
             return rc;
-        for (i = 0; i < n_handles; i++)
+        for (i = 0; (unsigned)i < n_handles; i++)
         {
             rc = cfg_get_inst_name(handles[i], &newiface);
             if (rc == 0 && strcmp(newiface, "lo") != 0)
